Update "get_locally_installed_packages" method
Currently this method is missing the "selenium" package during its detection, so I made some drastic changes to the method. Not sure if these are the sort of changes that I want to have for this project, but will review it later to make sure that I'm not missing other packages now.